Add files which is compatible with postgress database

+++ postgres-compatible/app.py
@@ -17,8 +17,12 @@ Session(app)
 
 # Creates a connection to the database
 #db = SQL ( "sqlite:///data.db" )
-# 
+
+# cloud heroku
 db = SQL ('postgresql://soqeugdursadgz:6870178bd4968172c42c8f58267a1188b2343cb48058de3848570af5d538e423@ec2-54-87-112-29.compute-1.amazonaws.com:5432/dcggcf0bnea4h5') # database engine object from SQLAlchemy that manages connections to the database
+
+#local
+#db = SQL ('postgresql://postgres:123456@localhost/snapcartt') # database engine object from SQLAlchemy that manages connections to the database
 
 @app.route("/")
 def index():
@@ -29,7 +33,7 @@ def index():
     shopLen = len(shoppingCart)
     totItems, total, display = 0, 0, 0
     if 'user' in session:
-        shoppingCart = db.execute("SELECT * FROM cart")
+        shoppingCart = db.execute("SELECT * FROM cart WHERE uid=:uid", uid=str(session["uid"]))
         shopLen = len(shoppingCart)
         for i in range(shopLen):
             total += shoppingCart[i]["subtotal"]
@@ -62,8 +66,8 @@ def buy():
         image = goods[0]["image"]
         subtotal = qty * price
         # Insert selected shirt into shopping cart
-        db.execute("INSERT INTO cart (id, qty, team, image, price, subtotal) VALUES (:id, :qty, :team, :image, :price, :subtotal)", id=id, qty=qty, team=team, image=image, price=price, subtotal=subtotal)
-        shoppingCart = db.execute("SELECT * FROM cart")
+        db.execute("INSERT INTO cart (id, qty, team, image, price, subtotal, uid) VALUES (:id, :qty, :team, :image, :price, :subtotal, :uid)", id=id, qty=qty, team=team, image=image, price=price, subtotal=subtotal, uid=str(session["uid"]))
+        shoppingCart = db.execute("SELECT * FROM cart WHERE uid=:uid", uid=str(session["uid"]))
         shopLen = len(shoppingCart)
         # Rebuild shopping cart
         for i in range(shopLen):
@@ -86,7 +90,7 @@ def update():
     if session:
         # Store id of the selected shirt
         id = int(request.args.get('id'))
-        db.execute("DELETE FROM cart WHERE id = :id", id=id)
+        db.execute("DELETE FROM cart WHERE id = :id and uid = :uid", id=id, uid=str(session["uid"]))
         # Select info of selected shirt from database
         goods = db.execute("SELECT * FROM shirts WHERE id = :id", id=id)
         # Extract values from selected shirt record
@@ -99,8 +103,8 @@ def update():
         image = goods[0]["image"]
         subtotal = qty * price
         # Insert selected shirt into shopping cart
-        db.execute("INSERT INTO cart (id, qty, team, image, price, subtotal) VALUES (:id, :qty, :team, :image, :price, :subtotal)", id=id, qty=qty, team=team, image=image, price=price, subtotal=subtotal)
-        shoppingCart = db.execute("SELECT * FROM cart")
+        db.execute("INSERT INTO cart (id, qty, team, image, price, subtotal, uid) VALUES (:id, :qty, :team, :image, :price, :subtotal, :uid)", id=id, qty=qty, team=team, image=image, price=price, subtotal=subtotal, uid=str(session["uid"]))
+        shoppingCart = db.execute("SELECT * FROM cart WHERE uid=:uid", uid=str(session["uid"]))
         shopLen = len(shoppingCart)
         # Rebuild shopping cart
         for i in range(shopLen):
@@ -134,7 +138,7 @@ def filter():
     totItems, total, display = 0, 0, 0
     if 'user' in session:
         # Rebuild shopping cart
-        shoppingCart = db.execute("SELECT * FROM cart")
+        shoppingCart = db.execute("SELECT * FROM cart WHERE uid=:uid", uid=str(session["uid"]))
         shopLen = len(shoppingCart)
         for i in range(shopLen):
             total += shoppingCart[i]["subtotal"]
@@ -147,12 +151,12 @@ def filter():
 
 @app.route("/checkout/")
 def checkout():
-    order = db.execute("SELECT * from cart")
+    order = db.execute("SELECT * from cart WHERE uid=:uid", uid=str(session["uid"]))
     # Update purchase history of current customer
     for item in order:
         db.execute("INSERT INTO purchases (uid, id, team, image, quantity) VALUES(:uid, :id, :team, :image, :quantity)", uid=str(session["uid"]), id=item["id"], team=item["team"], image=item["image"], quantity=item["qty"] )
     # Clear shopping cart
-    db.execute("DELETE from cart")
+    db.execute("DELETE from cart WHERE uid=:uid", uid=str(session["uid"]))
     shoppingCart = []
     shopLen = len(shoppingCart)
     totItems, total, display = 0, 0, 0
@@ -169,7 +173,7 @@ def remove():
     # Initialize shopping cart variables
     totItems, total, display = 0, 0, 0
     # Rebuild shopping cart
-    shoppingCart = db.execute("SELECT * FROM cart")
+    shoppingCart = db.execute("SELECT * FROM cart WHERE uid=:uid", uid=str(session["uid"]))
     shopLen = len(shoppingCart)
     for i in range(shopLen):
         total += shoppingCart[i]["subtotal"]
@@ -232,7 +236,7 @@ def history():
 @app.route("/logout/")
 def logout():
     # clear shopping cart
-    db.execute("DELETE from cart")
+    #db.execute("DELETE from cart WHERE uid=:uid", uid=str(session["uid"]))
     # Forget any user_id
     session.clear()
     # Redirect user to login form
@@ -266,7 +270,7 @@ def cart():
         # Clear shopping cart variables
         totItems, total, display = 0, 0, 0
         # Grab info currently in database
-        shoppingCart = db.execute("SELECT * FROM cart")
+        shoppingCart = db.execute("SELECT * FROM cart WHERE uid=:uid", uid=str(session["uid"]))
         # Get variable values
         shopLen = len(shoppingCart)
         for i in range(shopLen):
